+++ barcode/lib/barcode.dart
@@ -19,6 +19,7 @@ export 'src/barcode_exception.dart';
 export 'src/barcode_operations.dart';
 export 'src/barcode_types.dart';
 export 'src/codabar.dart' show BarcodeCodabarStartStop;
+export 'src/datamatrix.dart' show DataMatrixEncoder;
 export 'src/ean.dart' show BarcodeEan;
 export 'src/mecard.dart';
 export 'src/pdf417.dart' show Pdf417SecurityLevel;